short/PairProgrammingwithaLargeLanguageModel/requirements.txt
termcolor
google-api-python-client
google-generativeai
utils
